fix(frontend): use a native progress element and SSR-stable dates
Three findings from the scanners on 00990ec, all real.

Sonar typescript:S6819 wanted a native <progress> instead of a div carrying
role="progressbar". Switching to one is a straight improvement: the element
carries the semantics itself, so the hand-written aria-value* attributes go
away, and it clamps value to max on its own, which removes the percentage
arithmetic too. The fill is drawn by the browser, styled through
::-webkit-progress-value and ::-moz-progress-bar rather than an inner div.
Those two rules stay separate on purpose - an unrecognised pseudo element
invalidates a whole selector list, so combining them would make each engine
drop the rule the other needs.

react-doctor/no-locale-format-in-render on KeyTable: toLocaleDateString renders
with the server's locale and timezone during SSR and the reader's in the
browser, so the two passes disagree. Now formatted from toISOString, which is
UTC by definition and identical in both. That is also the better format for
this table - key metadata is developer-facing, where an unambiguous sortable
date beats one whose 03/04 you have to guess at. Invalid input falls back to the
same em dash as a null.

react-doctor/only-export-components on PlanCard: it exported PLANS and
BillingPlan alongside the component, which stops Fast Refresh from preserving
state. Both moved to plans.ts, which is where the copy belonged anyway.

Verified: 132 developer + sidebar tests pass, typecheck and lint clean, and the
browser QA over a fresh Storybook build still reports 11 stories with zero axe
violations. Screenshots checked by eye - the native progress renders the same
12% cyan bar and full-width red bar as the div it replaced.
